==> 12-Leer-datos-rds-INS.R <==
# 10-leer-datos-rds-INS.R

# Colombia
df.ins <- readRDS(file = "data/Colombia.RDS")

# Cali
df.ins.cali <- readRDS(file = "data/Cali.RDS")

# Buenaventura
df.ins.bventura <- readRDS(file = "data/Bventura.RDS")

# Palmira
df.ins.palmira <- readRDS(file = "data/Palmira.RDS")

# Jamundi
df.ins.jamundi <- readRDS(file = "data/Jamundi.RDS")

# Yumbo
df.ins.yumbo <- readRDS(file = "data/Yumbo.RDS")

# Tulua
df.ins.tulua <- readRDS(file = "data/Tulua.RDS")

# Valle del Cauca
df.ins.valle <- readRDS(file = "data/Valle.RDS")

# Valle Resto
df.ins.valle.resto <- readRDS(file = "data/Valle_resto.RDS")

# Valle No Distritos
df.ins.valle.nodis <- readRDS(file = "data/Valle_nodis.RDS")

# Medellin
df.ins.medellin <- readRDS(file = "data/Medellin.RDS")

# Bogota
df.ins.bogota <- readRDS(file = "data/Bogota.RDS")
