==> 20-Curva-Epidemica-Rt-Cali.Rmd <==
---
title: "Curva Epidémica y Número Efectivo de Reproducción (Rt) para COVID-19"
author: "Edgar Muñoz, Stat, MS Epid - munoze@outlook.com"
date: "5/25/2020"
output:
  html_document:
    df_print: paged
  pdf_document: default
  word_document: default
---

```{r configuracion, include = FALSE}
knitr::opts_chunk$set(echo = FALSE, message = FALSE, warning = FALSE)

source("00-Configuracion.R")
source("10-leer-datos-rds-INS.R")

# cargar las funciones
funciones <- list.files(path = "funciones", full.names = TRUE)
sapply(funciones, source)

# Personalizar paletas de colores para las curvas epidemicas
paleta_cf <- c(
  "Confirmados" = "orange",
  "Fallecidos"  = "blue"
)

paleta_pcf <- c(
  "Probables"   = "lightblue",
  "Confirmados" = "orange",
  "Fallecidos"  = "blue"
)

paleta_sexo <- c(
  "M" = "red",
  "F"  = "blue"
)

```


```{r prep-datos-area}

# fecha actual
hoy <- format(Sys.Date(), "%Y-%m-%d")

# definir el conjunto de analisis
df.casos <- readRDS("data/Cali.RDS")
area <- "Cali" # keep this label short

df.casos$persona <- 1.0

# fechas intervenciones
fecha_toquequeda  <- as.Date("2020-03-20")
fecha_cuarentena  <- as.Date("2020-03-25")
fecha_cuarentena1 <- as.Date("2020-04-13") # parcial cuarentena
fecha_cuarentena2 <- as.Date("2020-04-27") # parcial cuarentena
fecha_cuarentena3 <- as.Date("2020-05-11") # parcial cuarentena

```


```{r prep-inc-curva-epidemica, message=FALSE, warning=FALSE, include=FALSE}

# Fecha de notificacion mas reciente
per_fin_not <- max(df.casos$fecha_not)

# preparacion datos para Curva

# Incidencia - datos para analisis

# Fallecidos y Confirmados por fecha sintomas
inc.casos   <- incidence(df.casos$fecha_sin) 
max_ncasos  <- max(inc.casos$counts)
per_ini_sin <- as.Date(min(inc.casos$dates))
per_fin_sin <- as.Date(max(inc.casos$dates))

# Incidencia - datos para figura por tipo de caso
inc.casos.tipo <- incidence(df.casos$fecha_sin,  
                            groups = df.casos$tipo_caso)

# Eje x: 3 dias antes del 1er caso hasta 3 dias despues de hoy
xt_min = as.Date(per_ini_sin) - lubridate::days(3)
xt_max = as.Date(hoy) + lubridate::days(3)


```


```{r funcion-lineas-interv}

agregar_lineas_interv <- function(p)
{
  p <- p + geom_vline(xintercept = fecha_cuarentena, linetype = "dashed", color = "red") +
  annotate(
    "text",
    label = "Cuarentena",
    x = fecha_cuarentena,
    y = max_ncasos + 1,
    hjust = 1, angle = 90, vjust = -0.2
  ) +
  geom_vline(xintercept = fecha_toquequeda, linetype = "dashed", color = "red") +
  annotate(
    "text",
    label = "Toque de queda",
    x = fecha_toquequeda,
    y = max_ncasos + 1,
    hjust = 1, angle = 90, vjust = -0.2
  ) +
  geom_vline(xintercept = fecha_cuarentena1, linetype = "dashed", color = "red") +
  annotate(
    "text",
    label = "C. parcial",
    x = fecha_cuarentena1,
    y = max_ncasos + 1,
    hjust = 1, angle = 90, vjust = -0.2
  ) +
    geom_vline(xintercept = fecha_cuarentena2, linetype = "dashed", color = "red") +
  annotate(
    "text",
    label = "C. parcial",
    x = fecha_cuarentena2,
    y = max_ncasos + 1,
    hjust = 1, angle = 90, vjust = -0.2
  ) +
  geom_vline(xintercept = fecha_cuarentena3, linetype = "dashed", color = "red") +
  annotate(
    "text",
    label = "C. parcial",
    x = fecha_cuarentena3,
    y = max_ncasos + 1,
    hjust = 1, angle = 90, vjust = -0.2
  ) +
  geom_vline(xintercept = per_ini_sin, linetype = "dashed", color = "red") +
  annotate(
    "text",
    label = "1er caso",
    x = per_ini_sin,
    y = max_ncasos + 1,
    hjust = 1, angle = 90, vjust = -0.2
  ) +
  geom_vline(xintercept = as.Date(hoy), linetype = "dashed", color = "green") +
  annotate(
    "text",
    label = "Hoy",
    x = as.Date(hoy),
    y = max_ncasos + 1,
    hjust = 1, angle = 90, vjust = -0.2
  ) 
  p
}

```


```{r curva-epidemica-fn, message=FALSE, warning=FALSE, include=FALSE}

It_plot <- plot(inc.casos.tipo, stack = TRUE, border = "white") +
  scale_fill_manual(values = paleta_cf, limits = names(paleta_cf)) +
  labs(x = "Fecha de Síntomas", y = "Casos")  +
  theme_classic2() +
  scale_x_date(
    date_labels = "%b %d",
    date_breaks = "1 week",
    limits = c(xt_min, xt_max)
  ) +
  ggtitle(paste0("Curva Epidémica, COVID-19, ",area)) + 
  theme(
    legend.position = "top",
    legend.direction = "horizontal",
    legend.title = element_blank()
  )

It_plot <- agregar_lineas_interv(It_plot)

file_curva_pdf  <- 
   paste0("salidas/", hoy, "-",area,"-c19-curva-epidemica-final-", per_fin_not, ".pdf")
 
ggsave(file_curva_pdf, plot = It_plot, height = 8, width = 11 )

It_plot_final <- It_plot

It_plot

```



```{r simul-SI, message=FALSE, warning=FALSE, include=FALSE}

# parametros estimacion

mean_covid19_si <- 6.8 # 6.5
sd_covid19_si   <- 4.7 # 0.62

k = 3 # número de dias ponderados para estimar Rt en cada t

t_start <- seq(2, nrow(inc.casos)-(k-1))   
t_end <- t_start + (k-1)  

res_Rt_periodo <- estimate_R(inc.casos, method = "uncertain_si",
                  config = make_config(list(
                  mean_si = mean_covid19_si,
                  min_mean_si = 1.5, max_mean_si = 12.1, std_mean_si = 2.0, 
                  std_si = sd_covid19_si,
                  min_std_si = 1.5, max_std_si = 7.9, std_std_si = 2.0,
                  n1 = 100, n2 = 100,
                  t_start = t_start,
                  t_end = t_end)))

```


```{r grafica-SI, message=FALSE, warning=FALSE, include=FALSE}

plot_SI_periodo <- estimate_R_plots(res_Rt_periodo, what = "SI",
                            legend = FALSE)

plot_SI_periodo <- plot_SI_periodo + theme_pubclean() +
  labs(x = "Intervalo Serial (dias)", y= "Frecuencia")  +
  geom_vline(xintercept = mean_covid19_si , linetype="dashed", color = "orange" ) +
  annotate("text", label = paste0(" Media = ",mean_covid19_si, " dias"),
           x = mean_covid19_si, y = 0.8, hjust = 0) +
  ggtitle(paste0("Intervalo Serial (distribuciones simuladas)"))

file_curva_pdf  <- paste0("salidas/",hoy,"-", area, "-c19-distr-SI-per",".pdf")
ggsave(file_curva_pdf, plot = plot_SI_periodo, height = 8, width = 11 )

plot_SI_periodo

```


```{r funcion-intervenciones-rt}

agregar_lineas_interv_rt <- function(p)
{
  p <- p + geom_vline(xintercept = fecha_cuarentena, linetype = "dashed", color = "red") +
  annotate(
    "text",
    label = "Cuarentena",
    x = fecha_cuarentena,
    y = max_y,
    hjust = 1, angle = 90, vjust = -0.2
  ) +
  geom_vline(xintercept = fecha_toquequeda, linetype = "dashed", color = "red") +
  annotate(
    "text",
    label = "Toque de queda",
    x = fecha_toquequeda,
    y = max_y,
    hjust = 1, angle = 90, vjust = -0.2
  ) +
  geom_vline(xintercept = fecha_cuarentena1, linetype = "dashed", color = "red") +
  annotate(
    "text",
    label = "C. parcial",
    x = fecha_cuarentena1,
    y = max_y,
    hjust = 1, angle = 90, vjust = -0.2
  ) +
    geom_vline(xintercept = fecha_cuarentena2, linetype = "dashed", color = "red") +
  annotate(
    "text",
    label = "C. parcial",
    x = fecha_cuarentena2,
    y = max_y,
    hjust = 1, angle = 90, vjust = -0.2
  ) +
  geom_vline(xintercept = fecha_cuarentena3, linetype = "dashed", color = "red") +
  annotate(
    "text",
    label = "C. parcial",
    x = fecha_cuarentena3,
    y = max_y,
    hjust = 1, angle = 90, vjust = -0.2
  ) +
  geom_vline(xintercept = per_ini_sin, linetype = "dashed", color = "red") +
  annotate(
    "text",
    label = "1er caso",
    x = per_ini_sin,
    y = max_y,
    hjust = 1, angle = 90, vjust = -0.2
  ) +
  geom_vline(xintercept = as.Date(hoy), linetype = "dashed", color = "green") +
  annotate(
    "text",
    label = "Hoy",
    x = as.Date(hoy),
    y = max_y,
    hjust = 1, angle = 90, vjust = -0.2
  ) 
  p
}

```



```{r grafica-Rt-solo, include=FALSE}

plot_Rt_periodo <- estimate_R_plots(res_Rt_periodo, what = "R",
                            legend = FALSE, options_R = list(col = "blue")) 

# max Rt
data_Rt_per_all <- plot_Rt_periodo[["data"]]
max_y <- max(data_Rt_per_all[, "upper"], na.rm = TRUE) + 0.01

# continuar plotting 
plot_Rt_periodo <- plot_Rt_periodo + 
  labs(x = "", y = expression(R[t]) ) +
  scale_x_date(
    date_labels = "%b %d",
    date_breaks = "1 week",
    limits = c(xt_min, xt_max)
  ) +
  scale_y_continuous(
    breaks = c(0,1,seq(2, max_y, 2)),
    limits = c(0, max_y)
  ) +
  theme_classic2() +
  theme(
    legend.position = "top",
    legend.direction = "horizontal",
    legend.title = element_blank()
  ) +
  ggtitle(paste0("Número efectivo de reproducción (Rt), COVID-19, ", area))  

plot_Rt_periodo <- agregar_lineas_interv_rt(plot_Rt_periodo)

file_curva_pdf  <- paste0("salidas/",hoy, "-", area, "-c19-curva-Rt-per",".pdf")

ggsave(file_curva_pdf, plot = plot_Rt_periodo, height = 8, width = 11 )

plot_Rt_periodo

```


```{r grafica-Rt-It-pdf, message=FALSE, warning=FALSE, include=FALSE}

pg_Rt_It <- plot_Rt_It(plot_Rt_periodo, It_plot_final, ncol = 1)

file_curva_pdf  <- paste0("salidas/",hoy, "-", area, "-c19-curva-Rt-vs-It-", per_fin_not, ".pdf")

ggsave(file_curva_pdf, plot = pg_Rt_It, height = 8, width = 11 )

```

## Número Efectivo de Reproducción vs Curva Epidémica

```{r grafica-Rt-It, echo=FALSE, fig.height=9, message=FALSE, warning=FALSE, results='asis'}

plot_Rt_It(plot_Rt_periodo, It_plot_final, ncol = 1)

```


## Curva epidémica por sexo


```{r curva-epidemica-por-sexo}

# Generacion de la Curva - por Sexo

# Incidencia - datos para figura por sexo
inc.casos.sexo <- incidence(df.casos$fecha_sin,
                          groups = df.casos$sexo)

# Figura - por Sexo

It_plot_sexo <- plot(inc.casos.sexo, stack = TRUE, border = "white") +
  scale_fill_manual(values = paleta_sexo, limits = names(paleta_sexo)) +
  labs(x = "Fecha de Síntomas", y = "Casos")  +
  theme_classic2() +
  scale_x_date(
    date_labels = "%b %d",
    date_breaks = "1 week",
    limits = c(xt_min, xt_max)
  ) +
  ggtitle(paste0("Curva Epidémica, COVID-19, ", area)) + 
  theme(
    legend.position = "top",
    legend.direction = "horizontal",
    legend.title = element_blank()
  )

It_plot_sexo <- agregar_lineas_interv(It_plot_sexo)

file_curva_pdf  <- 
   paste0("salidas/", hoy, "-", area, "-c19-curva-epidemica-sexo-", per_fin_not, ".pdf")
 
ggsave(file_curva_pdf, plot = It_plot_sexo, height = 8, width = 11 )

It_plot_sexo

```


## Distribución de casos


```{r distribucion-casos}

df.casos %>%
  group_by(municipio) %>%
  summarise(
    casos = n()
    ) %>%
  mutate(porcentaje = round( (casos/sum(casos)*100),1)) %>%
  arrange(desc(casos)) -> df.municipios

kable(df.municipios, caption = "Distribución por municipio") %>%
  kable_styling(bootstrap_options = c("striped", "hover", "condensed", "responsive")) %>%
  scroll_box(width = "100%", height = "400px")

kable( with(df.casos, addmargins(table(tipo))),
      caption = "Distribución por tipo de caso importado/relacionado",
      col.names = c("Tipo de caso importado/relacionado", "n")) %>%
  kable_styling(bootstrap_options = c("striped", "hover", "condensed", "responsive"))

kable(with(df.casos, addmargins(table(atenci_n))),
      caption = "Distribución por tipo de atención médica (INS)",
      col.names = c("Atención médica", "n")) %>%
    kable_styling(bootstrap_options = c("striped", "hover", "condensed", "responsive"))

kable(with(df.casos, addmargins(table(estado))),
      caption = "Distribución por estado de salud (INS)",
      col.names = c("Estado de salud", "n")) %>%
    kable_styling(bootstrap_options = c("striped", "hover", "condensed", "responsive"))

kable(with(df.casos, addmargins(table(sexo))),
      caption = "Distribución por sexo",
      col.names = c("Sexo", "n")) %>%
    kable_styling(bootstrap_options = c("striped", "hover", "condensed", "responsive"))

kable(with(df.casos, addmargins(table(tipo_caso))),
      caption = "Distribución por tipo de caso confirmado",
      col.names = c("Tipo de caso confirmado/fallecido", "n")) %>%
    kable_styling(bootstrap_options = c("striped", "hover", "condensed", "responsive"))

```

Contacto: munoze@outlook.com
Tel. +57 300-929-1044


Colombia, 2020
